wallet
переделанный  swagger + app
