+++ middlewares/index.js
@@ -1,7 +1,10 @@
 const authenticate = require('./authenticate')
 const validate = require('./validate')
+const { validateUser, validateUserByEmail } = require('./utils')
 
 module.exports = {
   authenticate,
-  validate
+  validate,
+  validateUser,
+  validateUserByEmail
 }
