+++ package.json
@@ -28,7 +28,7 @@
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
     "swagger": "^0.7.5",
-    "swagger-jsdoc": "^6.1.0",
+    "swagger-jsdoc": "^5.0.1",
     "swagger-ui-express": "^4.1.6",
     "uuid": "^8.3.2"
   },
